requirements.txt
aniso8601==9.0.1
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.13.3
